requirements.txt
flask
numpy
opencv-python
scikit-image
scikit-learn
Pillow